refactor: move controller docs into v1/docs subpackage
