+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/api/controller/v1/docs/LocationControllerDocs.kt
@@ -1,4 +1,4 @@
-package noweekend.core.api.controller.v1
+package noweekend.core.api.controller.v1.docs
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
