+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/api/controller/v1/docs/MyPageControllerDocs.kt
@@ -1,27 +1,50 @@
-package noweekend.core.api.controller.v1
+package noweekend.core.api.controller.v1.docs
 
 import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.ExampleObject
 import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.parameters.RequestBody
-import noweekend.core.api.controller.v1.request.ScheduleCreateRequest
-import noweekend.core.api.controller.v1.request.ScheduleUpdateRequest
-import noweekend.core.api.controller.v1.response.DailyScheduleResponse
-import noweekend.core.api.controller.v1.response.ScheduleResponse
+import io.swagger.v3.oas.annotations.tags.Tag
+import noweekend.core.api.controller.v1.request.LeaveInputRequest
+import noweekend.core.api.controller.v1.request.ProfileRequest
+import noweekend.core.api.controller.v1.request.TagUpdateRequest
+import noweekend.core.api.security.annotations.CurrentUserId
+import noweekend.core.domain.tag.UserTags
 import noweekend.core.support.response.ApiResponse
-import java.time.LocalDate
 import io.swagger.v3.oas.annotations.responses.ApiResponse as SwaggerApiResponse
 
-interface ScheduleControllerDocs {
+@Tag(name = "마이페이지", description = "정보 수정 API")
+interface MyPageControllerDocs {
 
     @Operation(
-        summary = "캘린더: 일정 조회",
-        description = "startDate ~ endDate 구간에 포함된 DailySchedule 리스트를 반환합니다.",
+        summary = "마이페이지: 닉네임/생년월일 수정",
+        description = "유저가 마이페이지에서 닉네임과 생년월일을 수정합니다.",
+        requestBody = RequestBody(
+            required = true,
+            content = [
+                Content(
+                    mediaType = "application/json",
+                    schema = Schema(implementation = ProfileRequest::class),
+                    examples = [
+                        ExampleObject(
+                            name = "예시 요청",
+                            value = """
+                        {
+                          "nickname": "김매송..",
+                          "birthDate": "19991213"
+                        }
+                        """,
+                        ),
+                    ],
+                ),
+            ],
+        ),
         responses = [
             SwaggerApiResponse(
                 responseCode = "200",
-                description = "일정 조회 성공",
+                description = "프로필 수정 성공",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -32,24 +55,7 @@ interface ScheduleControllerDocs {
                                 value = """
 {
   "result": "SUCCESS",
-  "data": [
-    {
-      "date": "2025-05-01",
-      "schedules": [
-        {
-          "id": "abc123",
-          "title": "회의",
-          "startTime": "2025-05-01T10:00:00",
-          "endTime": "2025-05-01T11:00:00",
-          "category": "COMPANY",
-          "temperature": 3,
-          "allDay": false,
-          "alarmOption": "FIFTEEN_MINUTES_BEFORE",
-          "completed": false
-        }
-      ]
-    }
-  ],
+  "data": "프로필이 성공적으로 변경 되었습니다.",
   "error": null
 }
 """,
@@ -86,34 +92,209 @@ interface ScheduleControllerDocs {
             ),
         ],
     )
-    fun getSchedules(
-        @Schema(hidden = true) userId: String,
-        startDate: LocalDate,
-        endDate: LocalDate,
-    ): ApiResponse<List<DailyScheduleResponse>>
+    fun updateProfile(
+        @Parameter(hidden = true) @CurrentUserId userId: String,
+        profileRequest: ProfileRequest,
+    ): ApiResponse<String>
 
     @Operation(
-        summary = "캘린더: 일정 생성",
-        description = "일정을 생성합니다.",
+        summary = "마이페이지: 연차 정보 수정",
+        description = "유저가 연차 정보를 수정합니다.",
         requestBody = RequestBody(
             required = true,
             content = [
                 Content(
                     mediaType = "application/json",
-                    schema = Schema(implementation = ScheduleCreateRequest::class),
+                    schema = Schema(implementation = LeaveInputRequest::class),
+                    examples = [
+                        ExampleObject(
+                            name = "예시 요청",
+                            value = """
+                        {
+                          "days": 8,
+                          "hours": 4
+                        }
+                        """,
+                        ),
+                    ],
+                ),
+            ],
+        ),
+        responses = [
+            SwaggerApiResponse(
+                responseCode = "200",
+                description = "연차 정보 수정 성공",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "SUCCESS",
+  "data": "연차 정보가 성공적으로 변경되었습니다.",
+  "error": null
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+            SwaggerApiResponse(
+                responseCode = "400",
+                description = "잘못된 요청",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "ERROR",
+  "data": null,
+  "error": {
+    "code": "INVALID_PARAMETER",
+    "message": "올바르지 않은 요청입니다.",
+    "data": {}
+  }
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+        ],
+    )
+    fun updateLeave(
+        @Parameter(hidden = true) @CurrentUserId userId: String,
+        request: LeaveInputRequest,
+    ): ApiResponse<String>
+
+    @Operation(
+        summary = "마이페이지: 태그(자주하는 일정) 조회",
+        description = "유저의 선택/미선택 기본 태그와 커스텀 태그 목록을 반환합니다.",
+        responses = [
+            SwaggerApiResponse(
+                responseCode = "200",
+                description = "태그 조회 성공",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "SUCCESS",
+  "data": {
+    "selectedBasicTags": [
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-1234567890ab",
+            "content": "회의 참석",
+            "userId": "user-1",
+            "selected": true
+        },
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-2234567890ab",
+            "content": "점심 식사 약속",
+            "userId": "user-1",
+            "selected": true
+        }
+    ],
+    "unselectedBasicTags": [
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-2234567890ab",
+            "content": "헬스장 운동",
+            "userId": "user-1",
+            "selected": false
+        },
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-3234567890ab",
+            "content": "장 보기 / 마트 가기",
+            "userId": "user-1",
+            "selected": false
+        }
+    ],
+    "selectedCustomTags": [
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-4234567890ab",
+            "content": "스터디 그룹",
+            "userId": "user-1",
+            "selected": true
+        }
+    ],
+    "unselectedCustomTags": [
+        {
+            "id": "baf123e1-aaaa-bbbb-cccc-5234567890ab",
+            "content": "비즈니스 미팅",
+            "userId": "user-1",
+            "selected": false
+        }
+    ]
+  },
+  "error": null
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+            SwaggerApiResponse(
+                responseCode = "400",
+                description = "잘못된 요청",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "ERROR",
+  "data": null,
+  "error": {
+    "code": "INVALID_PARAMETER",
+    "message": "올바르지 않은 요청입니다.",
+    "data": {}
+  }
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+        ],
+    )
+    fun getTags(
+        @Parameter(hidden = true) @CurrentUserId userId: String,
+    ): ApiResponse<UserTags>
+
+    @Operation(
+        summary = "마이페이지: 태그(자주하는 일정) 수정",
+        description = "유저가 자신의 기본/커스텀 태그(자주하는 일정)를 추가·삭제(업서트)합니다.",
+        requestBody = RequestBody(
+            required = true,
+            content = [
+                Content(
+                    mediaType = "application/json",
+                    schema = Schema(implementation = TagUpdateRequest::class),
                     examples = [
                         ExampleObject(
                             name = "예시 요청",
                             value = """
 {
-  "title": "회의",
-  "date": "2025-05-01",
-  "startTime": "10:00:00",
-  "endTime": "11:00:00",
-  "category": "COMPANY",
-  "temperature": 3,
-  "allDay": false,
-  "alarmOption": "FIFTEEN_MINUTES_BEFORE"
+  "addScheduleTags": ["회의 참석", "헬스장 운동", "스터디 그룹"],
+  "deleteScheduleTags": ["장 보기 / 마트 가기", "비즈니스 미팅"]
 }
 """,
                         ),
@@ -124,7 +305,7 @@ interface ScheduleControllerDocs {
         responses = [
             SwaggerApiResponse(
                 responseCode = "200",
-                description = "일정 생성 성공",
+                description = "태그(자주하는 일정) 수정 성공",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -135,17 +316,7 @@ interface ScheduleControllerDocs {
                                 value = """
 {
   "result": "SUCCESS",
-  "data": {
-    "id": "abc123",
-    "title": "회의",
-    "startTime": "2025-05-01T10:00:00",
-    "endTime": "2025-05-01T11:00:00",
-    "category": "COMPANY",
-    "temperature": 3,
-    "allDay": false,
-    "alarmOption": "FIFTEEN_MINUTES_BEFORE",
-    "completed": false
-  },
+  "data": "자주하는 일정 태그가 성공적으로 수정되었습니다.",
   "error": null
 }
 """,
@@ -182,146 +353,8 @@ interface ScheduleControllerDocs {
             ),
         ],
     )
-    fun createSchedule(
-        @Schema(hidden = true) userId: String,
-        request: ScheduleCreateRequest,
-    ): ApiResponse<ScheduleResponse>
-
-    @Operation(
-        summary = "캘린더: 일정 수정",
-        description = "일정의 시작/종료 시간, 카테고리, 온도, 하루종일 여부, 알람 옵션을 수정합니다.",
-        requestBody = RequestBody(
-            required = true,
-            content = [
-                Content(
-                    mediaType = "application/json",
-                    schema = Schema(implementation = ScheduleUpdateRequest::class),
-                ),
-            ],
-        ),
-        responses = [
-            SwaggerApiResponse(
-                responseCode = "200",
-                description = "일정 수정 성공",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "SUCCESS",
-  "data": {
-    "id": "abc123",
-    "title": "회의",
-    "startTime": "2025-05-01T10:00:00",
-    "endTime": "2025-05-01T11:00:00",
-    "category": "COMPANY",
-    "temperature": 3,
-    "allDay": false,
-    "alarmOption": "FIFTEEN_MINUTES_BEFORE",
-    "completed": false
-  },
-  "error": null
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-            SwaggerApiResponse(
-                responseCode = "400",
-                description = "잘못된 요청",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "ERROR",
-  "data": null,
-  "error": {
-    "code": "INVALID_PARAMETER",
-    "message": "올바르지 않은 요청입니다.",
-    "data": {}
-  }
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-        ],
-    )
-    fun updateSchedule(
-        @Schema(hidden = true) userId: String,
-        id: String,
-        request: ScheduleUpdateRequest,
-    ): ApiResponse<ScheduleResponse>
-
-    @Operation(
-        summary = "캘린더: 일정 삭제",
-        description = "일정을 삭제합니다.",
-        responses = [
-            SwaggerApiResponse(
-                responseCode = "200",
-                description = "일정 삭제 성공",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "SUCCESS",
-  "data": "일정이 성공적으로 삭제되었습니다.",
-  "error": null
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-            SwaggerApiResponse(
-                responseCode = "400",
-                description = "잘못된 요청",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "ERROR",
-  "data": null,
-  "error": {
-    "code": "INVALID_PARAMETER",
-    "message": "올바르지 않은 요청입니다.",
-    "data": {}
-  }
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-        ],
-    )
-    fun deleteSchedule(
-        @Schema(hidden = true) userId: String,
-        id: String,
+    fun updateTags(
+        @Parameter(hidden = true) @CurrentUserId userId: String,
+        request: TagUpdateRequest,
     ): ApiResponse<String>
 }
