+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/api/controller/v1/docs/OnboardingControllerDocs.kt
@@ -1,4 +1,4 @@
-package noweekend.core.api.controller.v1
+package noweekend.core.api.controller.v1.docs
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
@@ -9,18 +9,96 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody
 import io.swagger.v3.oas.annotations.tags.Tag
 import noweekend.core.api.controller.v1.request.LeaveInputRequest
 import noweekend.core.api.controller.v1.request.ProfileRequest
-import noweekend.core.api.controller.v1.request.TagUpdateRequest
+import noweekend.core.api.controller.v1.request.TagRequest
+import noweekend.core.api.controller.v1.response.DefaultTags
 import noweekend.core.api.security.annotations.CurrentUserId
-import noweekend.core.domain.tag.UserTags
 import noweekend.core.support.response.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponse as SwaggerApiResponse
 
-@Tag(name = "마이페이지", description = "정보 수정 API")
-interface MyPageControllerDocs {
+@Tag(name = "온보딩", description = "온보딩 등록 API")
+interface OnboardingControllerDocs {
 
     @Operation(
-        summary = "마이페이지: 닉네임/생년월일 수정",
-        description = "유저가 마이페이지에서 닉네임과 생년월일을 수정합니다.",
+        summary = "온보딩: 자주하는 일정 등록",
+        description = "유저가 온보딩 시 자주하는 일정 태그(영문, Enum)를 리스트로 등록합니다.",
+        requestBody = RequestBody(
+            required = true,
+            content = [
+                Content(
+                    mediaType = "application/json",
+                    schema = Schema(implementation = TagRequest::class),
+                    examples = [
+                        ExampleObject(
+                            name = "예시 요청",
+                            value = """
+{
+  "scheduleTags": ["집안일", "은행 업무", "관공서 업무"]
+}
+""",
+                        ),
+                    ],
+                ),
+            ],
+        ),
+        responses = [
+            SwaggerApiResponse(
+                responseCode = "200",
+                description = "일정 등록 성공",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "SUCCESS",
+  "data": "일정 등록이 성공적으로 완료되었습니다.",
+  "error": null
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+            SwaggerApiResponse(
+                responseCode = "400",
+                description = "잘못된 요청",
+                content = [
+                    Content(
+                        mediaType = "application/json",
+                        schema = Schema(implementation = ApiResponse::class),
+                        examples = [
+                            ExampleObject(
+                                name = "예시 응답",
+                                value = """
+{
+  "result": "ERROR",
+  "data": null,
+  "error": {
+    "code": "INVALID_PARAMETER",
+    "message": "올바르지 않은 요청입니다.",
+    "data": {}
+  }
+}
+""",
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+        ],
+    )
+    fun registerSelectedDefaultTag(
+        @Parameter(hidden = true) @CurrentUserId userId: String,
+        request: TagRequest,
+    ): ApiResponse<String>
+
+    @Operation(
+        summary = "온보딩: 닉네임/생년월일 등록",
+        description = "유저가 온보딩 시 닉네임과 생년월일을 등록합니다.",
         requestBody = RequestBody(
             required = true,
             content = [
@@ -31,11 +109,11 @@ interface MyPageControllerDocs {
                         ExampleObject(
                             name = "예시 요청",
                             value = """
-                        {
-                          "nickname": "김매송..",
-                          "birthDate": "19991213"
-                        }
-                        """,
+{
+  "nickname": "김매송..",
+  "birthDate": "19991213"
+}
+""",
                         ),
                     ],
                 ),
@@ -44,7 +122,7 @@ interface MyPageControllerDocs {
         responses = [
             SwaggerApiResponse(
                 responseCode = "200",
-                description = "프로필 수정 성공",
+                description = "닉네임 및 생년월일 등록 성공",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -55,7 +133,7 @@ interface MyPageControllerDocs {
                                 value = """
 {
   "result": "SUCCESS",
-  "data": "프로필이 성공적으로 변경 되었습니다.",
+  "data": "닉네임 및 생년월일 등록이 성공적으로 완료되었습니다.",
   "error": null
 }
 """,
@@ -92,14 +170,14 @@ interface MyPageControllerDocs {
             ),
         ],
     )
-    fun updateProfile(
+    fun submitProfile(
         @Parameter(hidden = true) @CurrentUserId userId: String,
-        profileRequest: ProfileRequest,
+        request: ProfileRequest,
     ): ApiResponse<String>
 
     @Operation(
-        summary = "마이페이지: 연차 정보 수정",
-        description = "유저가 연차 정보를 수정합니다.",
+        summary = "온보딩: 올해 연차 입력",
+        description = "유저가 올해 남은 연차(일, 시간 단위)를 입력합니다.",
         requestBody = RequestBody(
             required = true,
             content = [
@@ -110,11 +188,11 @@ interface MyPageControllerDocs {
                         ExampleObject(
                             name = "예시 요청",
                             value = """
-                        {
-                          "days": 8,
-                          "hours": 4
-                        }
-                        """,
+{
+  "days": 5,
+  "hours": 4
+}
+""",
                         ),
                     ],
                 ),
@@ -123,7 +201,7 @@ interface MyPageControllerDocs {
         responses = [
             SwaggerApiResponse(
                 responseCode = "200",
-                description = "연차 정보 수정 성공",
+                description = "연차 등록 성공",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -134,7 +212,7 @@ interface MyPageControllerDocs {
                                 value = """
 {
   "result": "SUCCESS",
-  "data": "연차 정보가 성공적으로 변경되었습니다.",
+  "data": "연차 정보가 성공적으로 저장되었습니다.",
   "error": null
 }
 """,
@@ -171,18 +249,18 @@ interface MyPageControllerDocs {
             ),
         ],
     )
-    fun updateLeave(
+    fun submitLeave(
         @Parameter(hidden = true) @CurrentUserId userId: String,
         request: LeaveInputRequest,
     ): ApiResponse<String>
 
     @Operation(
-        summary = "마이페이지: 태그(자주하는 일정) 조회",
-        description = "유저의 선택/미선택 기본 태그와 커스텀 태그 목록을 반환합니다.",
+        summary = "온보딩: 기본 일정 태그 조회",
+        description = "온보딩 시 선택 가능한 기본 일정 태그(한글)를 모두 반환합니다.",
         responses = [
             SwaggerApiResponse(
                 responseCode = "200",
-                description = "태그 조회 성공",
+                description = "기본 일정 태그 조회 성공",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -194,49 +272,33 @@ interface MyPageControllerDocs {
 {
   "result": "SUCCESS",
   "data": {
-    "selectedBasicTags": [
-        {
-            "id": "baf123e1-aaaa-bbbb-cccc-1234567890ab",
-            "content": "회의 참석",
-            "userId": "user-1",
-            "selected": true
-        },
-        {
-            "id": "baf123e1-aaaa-bbbb-cccc-2234567890ab",
-            "content": "점심 식사 약속",
-            "userId": "user-1",
-            "selected": true
-        }
-    ],
-    "unselectedBasicTags": [
-        {
-            "id": "BASIC_TAG",
-            "content": "헬스장 운동",
-            "userId": "user-1",
-            "selected": false
-        },
-        {
-            "id": "baf123e1-aaaa-bbbb-cccc-3234567890ab",
-            "content": "장 보기 / 마트 가기",
-            "userId": "user-1",
-            "selected": false
-        }
-    ],
-    "selectedCustomTags": [
-        {
-            "id": "baf123e1-aaaa-bbbb-cccc-4234567890ab",
-            "content": "스터디 그룹",
-            "userId": "user-1",
-            "selected": true
-        }
-    ],
-    "unselectedCustomTags": [
-        {
-            "id": "baf123e1-aaaa-bbbb-cccc-5234567890ab",
-            "content": "비즈니스 미팅",
-            "userId": "user-1",
-            "selected": false
-        }
+    "tags": [
+      "회의 참석",
+      "점심 식사 약속",
+      "헬스장 운동",
+      "장 보기 / 마트 가기",
+      "가족 모임",
+      "병원 예약",
+      "카페에서 작업 / 휴식",
+      "친구 만남",
+      "술자리",
+      "스터디",
+      "학원 수업",
+      "야근",
+      "추가 업무",
+      "산책",
+      "반려동물 산책",
+      "집안일",
+      "은행 업무",
+      "관공서 업무",
+      "독서",
+      "데이트",
+      "미용실",
+      "드라이브",
+      "나들이",
+      "넷플릭스 시청",
+      "유튜브 시청",
+      "치지직 시청"
     ]
   },
   "error": null
@@ -275,86 +337,7 @@ interface MyPageControllerDocs {
             ),
         ],
     )
-    fun getTags(
+    fun getDefaultTag(
         @Parameter(hidden = true) @CurrentUserId userId: String,
-    ): ApiResponse<UserTags>
-
-    @Operation(
-        summary = "마이페이지: 태그(자주하는 일정) 수정",
-        description = "유저가 자신의 기본/커스텀 태그(자주하는 일정)를 추가·삭제(업서트)합니다.",
-        requestBody = RequestBody(
-            required = true,
-            content = [
-                Content(
-                    mediaType = "application/json",
-                    schema = Schema(implementation = TagUpdateRequest::class),
-                    examples = [
-                        ExampleObject(
-                            name = "예시 요청",
-                            value = """
-{
-  "addScheduleTags": ["회의 참석", "헬스장 운동", "스터디 그룹"],
-  "deleteScheduleTags": ["장 보기 / 마트 가기", "비즈니스 미팅"]
-}
-""",
-                        ),
-                    ],
-                ),
-            ],
-        ),
-        responses = [
-            SwaggerApiResponse(
-                responseCode = "200",
-                description = "태그(자주하는 일정) 수정 성공",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "SUCCESS",
-  "data": "자주하는 일정 태그가 성공적으로 수정되었습니다.",
-  "error": null
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-            SwaggerApiResponse(
-                responseCode = "400",
-                description = "잘못된 요청",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ApiResponse::class),
-                        examples = [
-                            ExampleObject(
-                                name = "예시 응답",
-                                value = """
-{
-  "result": "ERROR",
-  "data": null,
-  "error": {
-    "code": "INVALID_PARAMETER",
-    "message": "올바르지 않은 요청입니다.",
-    "data": {}
-  }
-}
-""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-        ],
-    )
-    fun updateTags(
-        @Parameter(hidden = true) @CurrentUserId userId: String,
-        request: TagUpdateRequest,
-    ): ApiResponse<String>
+    ): ApiResponse<DefaultTags>
 }
